mi_api/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
firebase-admin
pydantic